+++ src/content/mod.rs
@@ -4,9 +4,11 @@ mod section;
 mod sorting;
 mod utils;
 mod file_info;
+mod taxonomies;
 
 pub use self::page::{Page};
 pub use self::section::{Section};
 pub use self::pagination::{Paginator, Pager};
 pub use self::sorting::{SortBy, sort_pages, populate_previous_and_next_pages};
+pub use self::taxonomies::{Taxonomy, TaxonomyItem};