+++ src/content/page.rs
@@ -12,7 +12,7 @@ use errors::{Result, ResultExt};
 use config::Config;
 use front_matter::{PageFrontMatter, split_page_content};
 use markdown::markdown_to_html;
-use utils::{read_file};
+use fs::{read_file};
 use content::utils::{find_related_assets, get_reading_analytics};
 use content::file_info::FileInfo;
 
@@ -163,7 +163,7 @@ impl Default for Page {
 
 impl ser::Serialize for Page {
     fn serialize<S>(&self, serializer: S) -> StdResult<S::Ok, S::Error> where S: ser::Serializer {
-        let mut state = serializer.serialize_struct("page", 16)?;
+        let mut state = serializer.serialize_struct("page", 15)?;
         state.serialize_field("content", &self.content)?;
         state.serialize_field("title", &self.meta.title)?;
         state.serialize_field("description", &self.meta.description)?;
@@ -173,7 +173,6 @@ impl ser::Serialize for Page {
         state.serialize_field("permalink", &self.permalink)?;
         state.serialize_field("summary", &self.summary)?;
         state.serialize_field("tags", &self.meta.tags)?;
-        state.serialize_field("draft", &self.meta.draft)?;
         state.serialize_field("category", &self.meta.category)?;
         state.serialize_field("extra", &self.meta.extra)?;
         let (word_count, reading_time) = get_reading_analytics(&self.raw_content);