+++ src/content/section.rs
@@ -8,7 +8,7 @@ use serde::ser::{SerializeStruct, self};
 use config::Config;
 use front_matter::{SectionFrontMatter, split_section_content};
 use errors::{Result, ResultExt};
-use utils::{read_file};
+use fs::{read_file};
 use markdown::markdown_to_html;
 use content::Page;
 use content::file_info::FileInfo;